chore: remove helmet and used shared function to set og meta tags
