+++ packages/p2p/src/components/buy-sell/buy-sell.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Helmet } from 'react-helmet-async';
 import PropTypes from 'prop-types';
 import { observer } from 'mobx-react-lite';
+import { changeMetaTagWithOG } from '@deriv/shared';
 import { localize } from 'Components/i18next';
 import PageReturn from 'Components/page-return/page-return.jsx';
 import Verification from 'Components/verification/verification.jsx';
@@ -20,7 +20,19 @@ const BuySell = () => {
         const disposeAdvertIntervalReaction = buy_sell_store.registerAdvertIntervalReaction();
         buy_sell_store.setLocalCurrency(buy_sell_store.selected_local_currency);
 
+        const description_content = 'P2P Description';
+        const title_content = 'Deriv P2P';
+        const image =
+            'https://play-lh.googleusercontent.com/ah8RkaAnph2gouJ48fVeybeJgw-tu2dzTDYL7miccIWxvd0ZcK5-MM20bGxjpjb2lXU';
+
+        const restoreMetaTagWithOGDescription = changeMetaTagWithOG('description', description_content);
+        const restoreMetaTagWithOGTitle = changeMetaTagWithOG('title', title_content);
+        const restoreMetaTagWithOGImage = changeMetaTagWithOG('image', image);
+
         return () => {
+            restoreMetaTagWithOGDescription();
+            restoreMetaTagWithOGTitle();
+            restoreMetaTagWithOGImage();
             disposeIsListedReaction();
             disposeAdvertIntervalReaction();
         };
@@ -48,49 +60,6 @@ const BuySell = () => {
 
     return (
         <div className='buy-sell'>
-            <Helmet
-                meta={[
-                    {
-                        name: 'description',
-                        content: 'Helmet Description',
-                    },
-                    {
-                        name: 'og:title',
-                        content: 'Helmet Title',
-                    },
-                    {
-                        name: 'og:description',
-                        content: 'Helmet OG Description',
-                    },
-                    {
-                        name: 'og:image',
-                        content:
-                            'https://play-lh.googleusercontent.com/ah8RkaAnph2gouJ48fVeybeJgw-tu2dzTDYL7miccIWxvd0ZcK5-MM20bGxjpjb2lXU',
-                    },
-                    {
-                        name: 'twitter:card',
-                        content: 'Twitter Card',
-                    },
-                    {
-                        name: 'twitter:creator',
-                        content: 'Twitter Creator',
-                    },
-                    {
-                        name: 'twitter:title',
-                        content: 'Twitter Title',
-                    },
-                    {
-                        name: 'twitter:description',
-                        content: 'Twitter Description',
-                    },
-                    {
-                        name: 'referrer',
-                        content: 'origin',
-                    },
-                ]}
-            >
-                <title>Deriv P2P</title>
-            </Helmet>
             <BuySellHeader table_type={buy_sell_store.table_type} />
             <BuySellTable
                 key={buy_sell_store.table_type}
